refunding pending outgoing packets
