+++ x/delayedack/keeper/fraud_test.go
@@ -1,13 +1,18 @@
 package keeper_test
 
 import (
+	ibctransfer "github.com/cosmos/ibc-go/v6/modules/apps/transfer"
 	channeltypes "github.com/cosmos/ibc-go/v6/modules/core/04-channel/types"
 	commontypes "github.com/dymensionxyz/dymension/v3/x/common/types"
+	damodule "github.com/dymensionxyz/dymension/v3/x/delayedack"
 	dkeeper "github.com/dymensionxyz/dymension/v3/x/delayedack/keeper"
+	"github.com/dymensionxyz/dymension/v3/x/delayedack/types"
 )
 
 func (suite *DelayedAckTestSuite) TestHandleFraud() {
 	keeper, ctx := suite.App.DelayedAckKeeper, suite.Ctx
+	transferStack := damodule.NewIBCMiddleware(ibctransfer.NewIBCModule(suite.App.TransferKeeper), keeper)
+
 	rollappId := "testRollappId"
 	pkts := generatePackets(rollappId, 5)
 	rollappId2 := "testRollappId2"
@@ -32,7 +37,7 @@ func (suite *DelayedAckTestSuite) TestHandleFraud() {
 	_, err = keeper.UpdateRollappPacketWithStatus(ctx, pkts2[0], commontypes.Status_FINALIZED)
 	suite.Require().Nil(err)
 
-	err = keeper.HandleFraud(ctx, rollappId)
+	err = keeper.HandleFraud(ctx, rollappId, transferStack)
 	suite.Require().Nil(err)
 
 	suite.Require().Equal(0, len(keeper.ListRollappPackets(ctx, prefixPending1)))
@@ -41,6 +46,35 @@ func (suite *DelayedAckTestSuite) TestHandleFraud() {
 	suite.Require().Equal(1, len(keeper.ListRollappPackets(ctx, prefixFinalized)))
 	suite.Require().Equal(1, len(keeper.ListRollappPackets(ctx, prefixFinalized2)))
 }
+
+func (suite *DelayedAckTestSuite) TestDeletionOfRevertedPackets() {
+	keeper, ctx := suite.App.DelayedAckKeeper, suite.Ctx
+	transferStack := damodule.NewIBCMiddleware(ibctransfer.NewIBCModule(suite.App.TransferKeeper), keeper)
+
+	rollappId := "testRollappId"
+	pkts := generatePackets(rollappId, 5)
+	rollappId2 := "testRollappId2"
+	pkts2 := generatePackets(rollappId2, 5)
+
+	for _, pkt := range append(pkts, pkts2...) {
+		err := keeper.SetRollappPacket(ctx, pkt)
+		suite.Require().NoError(err)
+	}
+
+	err := keeper.HandleFraud(ctx, rollappId, transferStack)
+	suite.Require().Nil(err)
+
+	suite.Require().Equal(10, len(keeper.GetAllRollappPackets(ctx)))
+
+	keeper.SetParams(ctx, types.Params{EpochIdentifier: "minute"})
+	epochHooks := keeper.GetEpochHooks()
+	err = epochHooks.AfterEpochEnd(ctx, "minute", 1)
+	suite.Require().NoError(err)
+
+	suite.Require().Equal(5, len(keeper.GetAllRollappPackets(ctx)))
+}
+
+// TODO: test refunds of pending packets
 
 /* ---------------------------------- utils --------------------------------- */
